# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Royal Letters API - الرسائل الملكية
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface AuthResponse {
  authenticated: boolean;
  username?: string;
}

export interface Question {
  id: string;
  questionText: string;
  orderIndex: number;
}

export interface QuestionInput {
  questionText: string;
  answerText: string;
  orderIndex: number;
}

export type LetterLanguage =
  (typeof LetterLanguage)[keyof typeof LetterLanguage];

export const LetterLanguage = {
  arabic: "arabic",
  english: "english",
} as const;

export type LetterStatus = (typeof LetterStatus)[keyof typeof LetterStatus];

export const LetterStatus = {
  draft: "draft",
  sent: "sent",
  read: "read",
  replied: "replied",
} as const;

export interface Letter {
  id: string;
  title: string;
  recipientName: string;
  uniqueToken: string;
  isRead: boolean;
  readAt?: string | null;
  language: LetterLanguage;
  status: LetterStatus;
  createdAt: string;
  updatedAt: string;
}

export interface Reply {
  id: string;
  letterId: string;
  replyBody: string;
  replyFrom: string;
  createdAt: string;
}

export type LetterDetail = Letter & {
  body: string;
  questions: Question[];
  replies: Reply[];
};

export type CreateLetterRequestLanguage =
  (typeof CreateLetterRequestLanguage)[keyof typeof CreateLetterRequestLanguage];

export const CreateLetterRequestLanguage = {
  arabic: "arabic",
  english: "english",
} as const;

export type CreateLetterRequestStatus =
  (typeof CreateLetterRequestStatus)[keyof typeof CreateLetterRequestStatus];

export const CreateLetterRequestStatus = {
  draft: "draft",
  sent: "sent",
} as const;

export interface CreateLetterRequest {
  title: string;
  body: string;
  recipientName: string;
  language?: CreateLetterRequestLanguage;
  status?: CreateLetterRequestStatus;
  questions?: QuestionInput[];
}

export type UpdateLetterRequestLanguage =
  (typeof UpdateLetterRequestLanguage)[keyof typeof UpdateLetterRequestLanguage];

export const UpdateLetterRequestLanguage = {
  arabic: "arabic",
  english: "english",
} as const;

export type UpdateLetterRequestStatus =
  (typeof UpdateLetterRequestStatus)[keyof typeof UpdateLetterRequestStatus];

export const UpdateLetterRequestStatus = {
  draft: "draft",
  sent: "sent",
  read: "read",
  replied: "replied",
} as const;

export interface UpdateLetterRequest {
  title?: string;
  body?: string;
  recipientName?: string;
  language?: UpdateLetterRequestLanguage;
  status?: UpdateLetterRequestStatus;
  questions?: QuestionInput[];
}

export interface LetterResponse {
  letter: Letter;
}

export interface LetterDetailResponse {
  letter: LetterDetail;
}

export interface LetterListResponse {
  letters: Letter[];
  total: number;
}

export type LetterMetaResponseLanguage =
  (typeof LetterMetaResponseLanguage)[keyof typeof LetterMetaResponseLanguage];

export const LetterMetaResponseLanguage = {
  arabic: "arabic",
  english: "english",
} as const;

export type LetterMetaResponseStatus =
  (typeof LetterMetaResponseStatus)[keyof typeof LetterMetaResponseStatus];

export const LetterMetaResponseStatus = {
  draft: "draft",
  sent: "sent",
  read: "read",
  replied: "replied",
} as const;

export interface LetterMetaResponse {
  title: string;
  recipientName: string;
  language: LetterMetaResponseLanguage;
  status: LetterMetaResponseStatus;
  questions: Question[];
}

export type UnlockRequestAnswersItem = {
  questionId: string;
  answer: string;
};

export interface UnlockRequest {
  answers: UnlockRequestAnswersItem[];
}

export interface CreateReplyRequest {
  token: string;
  replyBody: string;
  replyFrom: string;
}

export interface ReplyResponse {
  reply: Reply;
}

export interface ReplyListResponse {
  replies: Reply[];
}

export type ListLettersParams = {
  status?: ListLettersStatus;
  search?: string;
};

export type ListLettersStatus =
  (typeof ListLettersStatus)[keyof typeof ListLettersStatus];

export const ListLettersStatus = {
  draft: "draft",
  sent: "sent",
  read: "read",
  replied: "replied",
} as const;